+++ src/boring_math/homepage/development.py
@@ -31,13 +31,16 @@ All non-typing related dependencies. Arrows point from modules to their dependen
         bgcolor="#957fb8";
         node [style=filled, fillcolor="#181616", fontcolor="#dcd7ba"];
         edge [color="#181616", fontcolor="#dcd7ba"];
-        integer_math -> "pythonic_fp.circulararray";
-        integer_math -> "pythonic_fp.iterables";
+        combinatorics -> "pythonic_fp.circulararray";
+        combinatorics -> "pythonic_fp.iterables";
+        combinatorics -> number_theory;
+        number_theory -> "pythonic_fp.circulararray";
+        number_theory -> "pythonic_fp.iterables";
         probability_distributions -> math;
         probability_distributions -> "mathplotlib.pyplot";
         probability_distributions -> "pythonic_fp.fptools";
         pythagorean_triples -> sys;
-        pythagorean_triples -> integer_math;
+        pythagorean_triples -> number_theory;
         recursive_functions -> sys;
     }