+++ src/boring_math/homepage/top_level_releases.py
@@ -7,10 +7,12 @@ These are the most resent *Boring Math* project PyPI releases.
 ======================================== ======================================= ========= =============
  PyPI and GitHub Name                    Python Module                           Version   Date Released
 ======================================== ======================================= ========= =============
- boring-math-integer-math [#]_           boring_math.integer_math                1.1.0     2025-10-TBD
- boring-math-probability-distributions   boring_math.probability_distributions   0.9.0     2025-10-TBD
- boring-math-pythagorean-triples         boring_math.pythagorean_triples         0.9.0     2025-10-TBD
- boring-math-recursive-functions         boring_math.recursive_functions         0.9.0     2025-10-TBD
+ boring-math-combinatorics               boring_math.combinatorics               2.0.0     2025-10-09
+ boring-math-integer-math [#]_           boring_math.integer_math                1.2.0     2025-10-09
+ boring-math-number-theory               boring_math.number_theory               2.0.0     2025-10-09
+ boring-math-probability-distributions   boring_math.probability_distributions   0.8.1     2025-08-04
+ boring-math-pythagorean-triples         boring_math.pythagorean_triples         0.8.2     2025-08-04
+ boring-math-recursive-functions         boring_math.recursive_functions         0.8.1     2025-08-04
  boring-math-special-functions           boring_math.special_functions           0.1.0     2025-10-TBD
 ======================================== ======================================= ========= =============
 
@@ -44,7 +46,7 @@ Remarks
 
 .. [#]
 
-    Maintainer soon plans to break up boring-math-integer-math
+    **DEPRECATED:** boring-math-integer-math was broken up
     into two PyPI projects, ``boring-math-number-theory`` and
     ``boring-math-combinatorics``.